requirements.txt
playsound == 1.2.2
pillow == 9.4.0